style(Reporte): Agregar padding superior a reportes
